refactor(llm): move parse_llm_method from CLI helpers to session resolver
  - Move parse_llm_method() from cli/llm_helpers.py to llm/session/resolver.py
  - Update imports in prompt.py, commit.py, and workflow files
  - Delete now-empty cli/llm_helpers.py
  - Function unchanged, only moved to correct architectural layer
