+++ src/mcp_coder/llm/session/resolver.py
@@ -1,4 +1,12 @@
-"""CLI utilities for LLM parameter handling."""
+"""Session management and LLM method resolution utilities.
+
+This module provides utilities for resolving session parameters and
+parsing LLM method strings into provider/method tuples.
+"""
+
+__all__ = [
+    "parse_llm_method",
+]
 
 
 def parse_llm_method(llm_method: str) -> tuple[str, str]:
@@ -9,9 +17,20 @@ def parse_llm_method(llm_method: str) -> tuple[str, str]:
 
     Returns:
         Tuple of (provider, method)
+        - provider: "claude"
+        - method: "cli" or "api"
 
     Raises:
         ValueError: If llm_method is not supported
+
+    Example:
+        >>> provider, method = parse_llm_method("claude_code_api")
+        >>> print(provider, method)
+        claude api
+
+        >>> provider, method = parse_llm_method("claude_code_cli")
+        >>> print(provider, method)
+        claude cli
     """
     if llm_method == "claude_code_cli":
         return "claude", "cli"
